requirements.txt
pygame
pathfinding
pillow